bump reth to v1.0.8 (#1427)
* bump reth to v1.0.8

* fix

* fmt

* fix tests

* trunk fmt

* fix tests

* fix features

* fix clippy

* cargo update

* small fix

* some fix

* fix new reth rev after address serialization patch

* cargo update

* fix conflicts

* update submodule

+++ src/providers/eth_provider/constant.rs
@@ -1,7 +1,6 @@
-use std::{str::FromStr, sync::LazyLock};
-
-use reth_primitives::{B256, U256};
+use alloy_primitives::{B256, U256};
 use serde::{Deserialize, Serialize};
+use std::{str::FromStr, sync::LazyLock};
 
 /// Maximum priority fee per gas
 pub static MAX_PRIORITY_FEE_PER_GAS: LazyLock<u64> = LazyLock::new(|| 0);
@@ -11,7 +10,7 @@ pub static MAX_LOGS: LazyLock<Option<u64>> =
     LazyLock::new(|| std::env::var("MAX_LOGS").ok().and_then(|val| u64::from_str(&val).ok()));
 
 /// Gas limit for estimate gas and call
-pub const CALL_REQUEST_GAS_LIMIT: u128 = 50_000_000;
+pub const CALL_REQUEST_GAS_LIMIT: u64 = 50_000_000;
 /// Number of characters for representing a U256 in a hex string form. Used for padding hashes
 pub const HASH_HEX_STRING_LEN: usize = 64;
 /// Number of characters for representing logs topics in a hex string form. Used for padding logs topics
